--- a/CHANGES.txt
+++ b/CHANGES.txt
@@ -16,11 +16,12 @@ RELEASE  VERSION/DATE TO BE FILLED IN LATER
     - Whatever John Doe did.
 
   From Joseph Brill:
-    - The default Windows powershell 7 path is added before the default
-      Windows powershell 5 path in the limited environment in which the
-      MSVC batch files are run.  This appears to fix a runtime issue in
-      the windows runner environment when the embedded Visual Studio vcpkg
-      component initializes without a PSModulePath defined. The OS
+    - MSVS: Fix a significant MSVC/MSVC tool initialization slowdown when
+      vcpkg has been installed and the PSModulePath is not initialized
+      or propagated from the user's shell environment. To resolve this
+      The default Windows Powershell 7 path is added before the default
+      Windows Powershell 5 path in the carefully constructed
+      environment in which the MSVC batch files are run.   The shell
       environment variables and values for VCPKG_DISABLE_METRICS and
       VCPKG_ROOT are added to the limited MSVC environment when defined.
       At present, the VCPKG_DISABLE_METRICS and VCPKG_ROOT variables and