--- a/RELEASE.txt
+++ b/RELEASE.txt
@@ -35,6 +35,10 @@ CHANGED/ENHANCED EXISTING FUNCTIONALITY
   one from PEP 308 introduced in Python 2.5 (2006). The idiom being
   replaced (using and/or) is regarded as error prone.
 
+- The default Windows powershell 7 path is added before the default
+  Windows powershell 5 path in the limited environment in which the
+  MSVC batch files are run.
+
 FIXES
 -----