--- a/SCons/Tool/MSCommon/common.py
+++ b/SCons/Tool/MSCommon/common.py
@@ -354,7 +354,6 @@ def normalize_env(env, keys, force: bool=False):
     if progfiles_ps_dir not in normenv["PATH"]:
         normenv["PATH"] = normenv["PATH"] + os.pathsep + progfiles_ps_dir
 
-    # Powershell 5
     # Without Powershell in PATH, an internal call to a telemetry
     # function (starting with a VS2019 update) can fail
     # Note can also set VSCMD_SKIP_SENDTELEMETRY to avoid this.
@@ -362,19 +361,7 @@ def normalize_env(env, keys, force: bool=False):
     if sys32_ps_dir not in normenv['PATH']:
         normenv['PATH'] = normenv['PATH'] + os.pathsep + sys32_ps_dir
 
-    psmodulepath_dirs = [
-        # Powershell 7 paths
-        os.path.join(progfiles_dir, r"PowerShell\Modules"),
-        os.path.join(progfiles_dir, r"PowerShell\7\Modules"),
-        # Powershell 5 paths
-        os.path.join(progfiles_dir, r"WindowsPowerShell\Modules"),
-        os.path.join(sys32_dir, r"WindowsPowerShell\v1.0\Modules"),
-    ]
-
-    normenv["PSModulePath"] = os.pathsep.join(psmodulepath_dirs)
-
     debug("PATH: %s", normenv['PATH'])
-    debug("PSModulePath: %s", normenv['PSModulePath'])
     return normenv
 
 
@@ -411,11 +398,11 @@ def get_output(vcbat, args=None, env=None, skip_sendtelemetry=False):
         'VS71COMNTOOLS',
         'VSCOMNTOOLS',
         'MSDevDir',
-        'VCPKG_DISABLE_METRICS',
-        'VCPKG_ROOT',
         'VSCMD_DEBUG',   # enable logging and other debug aids
         'VSCMD_SKIP_SENDTELEMETRY',
         'windir', # windows directory (SystemRoot not available in 95/98/ME)
+        'VCPKG_DISABLE_METRICS',
+        'VCPKG_ROOT',
     ]
     env['ENV'] = normalize_env(env['ENV'], vs_vc_vars, force=False)