Fix for vcpkg run time in windows runner (VS2022 Enterprise).

Changes to msvc batch file invocation environment:
* Add Powershell 7 to PATH
* Import VCPKG_DISABLE_METRICS if defined
* Import VCPKG_ROOT if defined

## Contributor Checklist:

* [ ] I have created a new test or updated the unit tests to cover the new/changed functionality.
* [x] I have updated `CHANGES.txt` and `RELEASE.txt` (and read the `README.rst`).
* [ ] I have updated the appropriate documentation