@jcbrill 2025-05-07
@mwichmann The VS2017/PY37 environment is reporting that the MSVS 2015 license has expired.

Image: Visual Studio 2017; Environment: WINPYTHON=Python37 Test failure:
```
720/1305 ( 55.17%) C:\Python37\python.exe test\MSVS\vs-14.0-exec.py
C:\projects\scons\scripts\scons.py returned 1
STDOUT =========================================================================
Microsoft Visual Studio 2015 Version 14.0.25420.1.
Copyright (C) Microsoft Corp. All rights reserved.
The license for Visual Studio has expired.
Your license for this product is expired and could not be refreshed. Please connect to the internet to continue to use this product.
STDERR =========================================================================
FAILED test of C:\projects\scons\scripts\scons.py
	at line 753 of C:\projects\scons\testing\framework\TestCommon.py (_complete)
	from line 867 of C:\projects\scons\testing\framework\TestCommon.py (run)
	from line 485 of C:\projects\scons\testing\framework\TestSCons.py (run)
	from line 941 of C:\projects\scons\testing\framework\TestSConsMSVS.py (run)
	from line 104 of test\MSVS\vs-14.0-exec.py
Test execution time: 9.0 seconds
```