@mwichmann 2025-05-07
yes. @bdbaddog has contacted them about that (a couple of weeks ago, I think). they were supposed to be fixing.