@jcbrill 2025-05-07
> does this relate at all to #4226 ?

I'm going to pretend like I understood the details of #4226 and answer confidently with "only tangentially?"

My understanding of #4226, is the intent is to use vcpkg via the "standalone" vcpkg installation.

The windows runner contains two `vcpkg` installations:
1) a standalone vcpkg installation (C:\vcpkg and VCPKG_INSTALLATION_ROOT=C:\vcpkg)
2) an embedded vcpkg installation in the VS2022 Enterprise tree

Note: it ***may** be possible to have the both installations use the same package root.  This needs to be investigated.

Additional VS components may put initialization batch files in the `<VSROOT>\Common7\Tools\vsdevcmd\ext` directory.  The `vcvarsall.bat` calls `VsDevCmd.bat` which in turn calls all of the batch files in the `ext` folder.

A simplified view of the VS directory contents for vcpkg is:
* `<VSROOT>\Common7\Tools\vsdevcmd\ext`:
  * `vcpkg.bat`
* `<VSROOT>\VC\vcpkg`:
  * `vcpkg.exe`
  * `vcpkg-init.cmd`

The initial call sequence is to setup the embedded VS vcpkg installation which is done using powershell and may download from the GitHub vcpkg repository.

The embedded VS vcpkg batch files and powershell scripts run within the limited SCons msvc environment as it is kicked off from the msvc batch files.

The windows runner environment has both powershell 7 and powershell 5 installed (different locations so side-by-side).  The current SCons implementation adds powershell 5 to msvc environment PATH (i.e., `env['ENV']['PATH']` but not powershell 7.  In addition, the msvc environment does not contain a `PSModulePath` entry.  PSModulePath is the powershell search path for modules analogeous to the system path.

In the windows runner, adding powershell 7 before powershell 5 in the msvc environment system path and adding a `PSModulePath` entry containing the All users and system locations for powershell 7 and the All users and system locations for powershell 5 seems to have resolved the runtime issue.  My guess is that the newer powershell 7 versions work more effectively than the powershell 5 versions in this context.

The first call tended to be more expensive than later calls.  There was something about using powershell 5 without a PSModulePath that caused a problem.  The environment in which vcpkg.exe was executed does not contain Windows system environment variables (e.g., ProgramFiles, ProgramData, PROCESSOR_ARCHITECTURE, etc).

It is kind of difficult to debug the msvc batch file and powershell call chain as it is launched via a subprocess and there is limited diagnostic information available.

It may be possible to use the "standalone" package root with the embedded vcpkg.exe by setting `VCPKG_ROOT` to the standalone package root (e.g., VCPKG_ROOT=C:\vcpkg) in the system environment and "importing" the value into the msvc environment before calling the msvc batch files.

This PR passes along `VCPKG_ROOT` if it is defined.  Note however, the standalone vcpkg appears to use `VCPKG_INSTALLATION_ROOT` for the location.  Effectively doing something like `set VCPKG_ROOT=%VCPKG_INSTALLATION_ROOT%` ***might*** configure the embedded VS vcpkg to use the standalone vcpkg folder.  This almost seems like a bad idea.

This PR is intended to be the minimum required to workaround the initialization penalty for the embedded vcpkg.

I am planning on trying a few things locally.  Unfortunately, a local instance is not ephemeral like the runner image which makes alternate configurations and testing a bit more work.

Clear as mud?