@mwichmann 2025-05-07
So is the summary of the summary that things don't work well if `vcpkg` is used?  There was no `vcpkg` involved in the quickie "examples" build, and there was still that surprising extra minute.