@jcbrill 2025-05-07
Vcpkg isn't even being "used" it was simply installed as a component of Visual Studio.

The vcvars64.bat batch file is *initializing* the vcpkg component environment and experiences delays due to the aggressively limited environment in which the msvc batch files are run.  

I believe that the in-tree vcpkg data after initialization is lost when the runner completes so there is a runtime penalty every time the runner is invoked and to a lesser extend any subsequence invocations in the same session.

The persistent SCons MSVC cache in the windows runner environment limits the penalty to the first invocation of each combination of msvc batch file and arguments.  If the msvc cache is not persisted between runs, the runtime penalty applies every time.

What I can't answer is:
* is the vcpkg component new in the vs2022 runner?
* is powershell 7 new in the vs2022 runner?