@mwichmann 2025-05-07
My last question crossed paths, nevermind.

A fully executed "Developer Command Prompt" (that is, all the setup scripts executed) has a LOT of extra variables, which we knew, we were largely guessing at the minimal set.  Certainly has a `VCPKG_ROOT`. That's not new - I see an installed standalone `vcpkg` in all three current images, and a `Microsoft.VisualStudio.Component.Vcpkg` in the 2022 and 2025 images.  The available information is here:
```
[Microsoft.VisualStudio.Component.Vcpkg](https://github.com/actions/runner-images/blob/main/images/windows/Windows2025-Readme.md)
```

There's a `PSModulePath` in any windows shell (cmd or ps) I open on a Win11 Pro box. Presumably that only matters if powersehll-y things are done.