@jcbrill 2025-05-08
*I have reason to believe that simply adding powershell 7 to the msvc environment path in front of powershell 5 is the solution and that adding PSModulePath is not necessary.*

I ***believe*** when a powershell binary is invoked, if `PSModulePath` is not defined, the binary will search default locations for modules.

Notes:
* Without powershell 7 on the path, powershell 5 is using the powershell 5 modules which appear to be "slower".
* With powershell 7 on the path, powershell 7 is using the powershell 7 modules which appear to be "faster".
* If powershell 7 is not installed, the binary will not be found on the path and powershell 5 is the only alternative.
* Adding `PSModulePath` without adding powershell 7 to the path seemed to work as powershell 5 is likely using the powershell 7 modules.

`VCPKG_DISABLE_METRICS` is added to the msvc environment if it is defined.
`VCPKG_ROOT` is added to the msvc environment if it is defined.

Sample output from windows runner 2022 test environment (PATH with PS7, No PSModulePATH):
```
00524ms:SConstruct:_test_scons#1252: env[SystemDrive]=C:
00524ms:SConstruct:_test_scons#1252: env[SystemRoot]=C:\Windows
00524ms:SConstruct:_test_scons#1252: env[TEMP]=C:\Users\RUNNER~1\AppData\Local\Temp
00524ms:SConstruct:_test_scons#1252: env[TMP]=C:\Users\RUNNER~1\AppData\Local\Temp
00524ms:SConstruct:_test_scons#1252: env[USERPROFILE]=C:\Users\runneradmin
00524ms:SConstruct:_test_scons#1252: env[ComSpec]=C:\Windows\system32\cmd.exe
00524ms:SConstruct:_test_scons#1252: env[OS]=Windows_NT
00524ms:SConstruct:_test_scons#1252: env[VSCMD_SKIP_SENDTELEMETRY]=1
00524ms:SConstruct:_test_scons#1252: env[windir]=C:\Windows
00524ms:SConstruct:_test_scons#1252: env[VCPKG_DISABLE_METRICS]=1
00524ms:SConstruct:_test_scons#1252: env[PATH]=C:\Windows\System32;C:\Windows\System32\Wbem;C:\Program Files\PowerShell\7;C:\Windows\System32\WindowsPowerShell\v1.0
00524ms:SConstruct:_test_scons#1252: env[PATHEXT]=.COM;.EXE;.BAT;.CMD
03086ms:SConstruct:msvc_find_valid_batch_script#935: ELAPSED_TIME=2.56, vc_script='C:\\Program Files\\Microsoft Visual Studio\\2022\\Enterprise\\VC\\Auxiliary\\Build\\vcvars64.bat'
04197ms:SConstruct:msvc_find_valid_batch_script#935: ELAPSED_TIME=1.11, vc_script='C:\\Program Files\\Microsoft Visual Studio\\2022\\Enterprise\\VC\\Auxiliary\\Build\\vcvars64.bat'
05228ms:SConstruct:msvc_find_valid_batch_script#935: ELAPSED_TIME=1.03, vc_script='C:\\Program Files\\Microsoft Visual Studio\\2022\\Enterprise\\VC\\Auxiliary\\Build\\vcvars64.bat'
06245ms:SConstruct:msvc_find_valid_batch_script#935: ELAPSED_TIME=1.02, vc_script='C:\\Program Files\\Microsoft Visual Studio\\2022\\Enterprise\\VC\\Auxiliary\\Build\\vcvars64.bat'
07292ms:SConstruct:msvc_find_valid_batch_script#935: ELAPSED_TIME=1.05, vc_script='C:\\Program Files\\Microsoft Visual Studio\\2022\\Enterprise\\VC\\Auxiliary\\Build\\vcvars64.bat'
```